Get rid of the 'ggwp-' prefix for zgui and zscene
