+++ zgui/examples/nested.rs
@@ -1,10 +1,10 @@
 use cgmath::Point2;
-use ggwp_zgui as ui;
 use gwg::{
     conf, event,
     graphics::{self, Font, Image, Rect, Text},
     Context, GameResult,
 };
+use zgui as ui;
 
 #[derive(Clone, Copy, Debug)]
 enum Message {
