+++ zgui/resources/fire.png
