+++ zscene/README.md
@@ -1,6 +1,6 @@
-# `ggwp-zscene`
+# `zscene`
 
-`ggwp-zscene` is a simple scene and declarative animation manager.
+`zscene` is a simple scene and declarative animation manager.
 
 Made for Zemeroth game.
 Historically was part of [Häte](https://docs.rs/hate) crate.
@@ -29,11 +29,3 @@ self.scene.add_action(action.boxed());
 ```
 
 See [examples/action.rs](./examples/action.rs) for a complete example.
-
-## What does a `ggwp-` prefix mean?
-
-As Icefoxen asked to [not use `ggez-` prefix][ggwp]
-I use `ggwp-` ("good game, well played!") to denote that the crate
-belongs to ggez ecosystem, but is not official.
-
-[ggwp]: https://github.com/ggez/ggez/issues/373#issuecomment-390461696
