+++ zscene/resources/fire.png
